SI-6028 Restore original parameter names after lambdalift
Extends what was being done only in the pickled signatures to
the bytecode proper. Once the code has been lifted out of a tree,
we can undo the mangling in that tree.

This only does this for parameter name, but I suspect it could and
should be done for all symbols, which will present a few less mangled
 names in the debugger.

Do we have any test infrastructure for verifying this? Can we compare
-Xprint:lambdalift against a checkfile?