SI-6358 Fix lazy vals in value classes
The bug seems to be in changeOwner, rather than in value
classes themselves. It wasn't changing the owner of
`TermSymbol#referenced`, which is used to store the original
symbol `x` that is replaced by `x$lzy` in the lazyvals phase.

In ExtensionMethods, I threw in a `duplicate` of the method body
for good measure before it was grafted onto the companion object.
I don't have a failing test to motivate that change, but it is
in line with the similar code in `AddInterfaces#implMethodDef`.

It will be interesting to test all of this under the resident
compiler, to check we don't run into the likes of SI-5167.

It seems likely that other symbol oriented traversals are prone
to the same flaw.