SI-6624 Workaround poorly inferred pattern type
In this bug, we see an inferred type of the form:

  class Base[M[_]]; case class Derived[M[_]]
  type Dodgy = Base[Option] with Derived[Option]

This seems to stem from a problem glb-ing the parent
types when the type argument for one is still an unsolved
HK type variable.

The redundancy of the type turns into a material
problem when the pattern matcher tries to find the
case field accessors: it looks through the declarations
of the refined types and comes up with bubkus. We
need to look in Derived.

This patch is just a workaround, it doesn't address
the root problem. We re-run the glb after the type
vars have been solved.

review by @adriaanm