SI-6879 Don't inline methods calling Class.forName, etc.
The etcetera is the hard part. Inlining can change the meaning
of code that reflects on the call stack to find the callers
class or class loader.

This commit prevents inlining of methods that themselves call
a catalog of methods in the Java standard library, which are
known to use these tricks.

A less tractable problem managing the interaction with
SecurityManager. Here, it's not just the directly enclosing
stack frame that counts, but the entire call stack. Inlining
might remove a critical frame that was granted some permission.
The majority of applications don't run under such fine-grained
security policies. But those that do may need to disable
inlining in some cases.