SI-5082 Brandish a machete in the name of cycle avoidance.
The previous commit attempted to break a cycle triggered
by case class unapply synthesis. But in doing so, it is
unable to find the already-renamed case accessors for
non-public constructor parameters, have already been
assigned `$N` suffixed names in `caseTemplateBody`.

This leaves is in the woods, and we have to hack our way out.
Inspired by similar feats in setting up the case class copy
method, we install a custom type completer on the unapply
method, which patches the RHS to point at the correct
tree accessors.

This seems too ugly and brittle to propose for inclusion,
but points again to a need a little more laziness in DefDef
creation. (Other examples: using explicit type arguments in
Tuple.apply in the RHS of the unapply method to avoid SIP-18
non-wildcard existential warnings, and the copy method.)