SI-6900 Fix tailrec for dependent method types.
Uncurry's info transformer could generate a MethodType
with cloned parameter symbols. This type was used
for the LabelDef generated in the TailCalls phase.
But, the RHS of the method still contains types that
refer to the original parmameter symbol. Spurious,
type errors ensued.

I can't see my way to a principled fix, but would
love to hear any ideas for one. For now, I've cast
the arguments to and the result of the label jump
generated in TailCalls.