WIP SI-5465 Make Unapply trees retypeable through macros.
Typechecking of a pattern match that uses extractors resolves
a particular unapply method and rewrites the tree to refer to this.

The two trees below show the pre- and post- typechecking state of
this snippet:

    object Extractor { def unapply(x: Int): Option[Int] = Some(x) }
    val extractor = reify {
      2 match { case Extractor(x) => x }
    }

Trees:

    Match(Literal(Constant(2)), List(CaseDef(        Apply(Select(This(newTypeName("Test")), newTermName("Extractor")),                                                                           List(Bind(newTermName("x"), Ident(nme.WILDCARD)))), EmptyTree, Ident(newTermName("x")))))
    Match(Literal(Constant(2)), List(CaseDef(UnApply(Apply(Select(Select(This(newTypeName("Test")), Test.Extractor   ), newTermName("unapply")), List(Ident(newTermName("<unapply-selector>")))), List(Bind(newTermName("x"), Ident(nme.WILDCARD)))), EmptyTree, Ident(newTermName("x")))))

The resulting tree is not amenable to a `resetAttrs` and a re-typecheck.

This commit changes `ResetAttrs` to revert this transformation.

With this patch, we can satisfy the test case from SI-7230 (using macros),
but we are unable to fix the existing test, run/idempotency-extractors.scala,
(which uses the toolbox compiler.) This is reminiscent of the fix for SI-6187,
which had the same limitation.