SI-7482 Fix regression in raw type cookery
In 9e101a3d, the cooking of raw types was refactored.
Among the changes was a seemingly innocuous use of `modifyInfo`.
But this unconditionally adds a new entry in the type history,
even if the function (in this case, `rawToExistential) doesn't
modify the type.

For reasons that are still not 100% clear, that leads to problems
in a subsequent compilation run, at least in the REPL. In the
second run, the info of the the constructor of, say `ArrayList`
has lost track of its origins, and contains an existential result
type:

   TypeHistory(typer:7,(x$1: Int)java.util.ArrayList[_],null)

Comparing with another constructor on that class:

  TypeHistory(posterasure:5,(x$1: java.util.Collection)java.util.ArrayList,
    TypeHistory(namer:7,(x$1: java.util.Collection[_ <: E])java.util.ArrayList[E],null))