SI-7518 Better positioning for inlined code
The inliner was positioning all inlined instructions
at the method call (`targetPos). This makes debugging
of code compiled under `-optimize` difficult.

This commit retains the original position of the
inlined instruction if it is in the same source file
as the method call. In other cases, we have to fall
back to the `targetPos`; line number information in
classfiles can only index into current source file.