Move ICodeReader-specific logic out of ClassfileParser.
ClassfileParser contained some ICodeReader-specific logic
like `forceMangledName` and `getMemberSymbol` methods.

The `getMemberSymbol` method was defined in ConstantPool class
because it must access some internal state of ConstantPool. In order
to move that method to ICodeReader we had two options:

  1. Make all internal state accessible from outside of ConstantPool
     class so getMemberSymbol could be implemented outside of
     ConstantPool hierarchy
  2. Make it possible to subclass ConstantPool in ICodeReader so
     getMemberSymbol can be implemented in a subclass and can access
     internal state of ConstantPool

Given the fact that getMemberSymbol mutates ConstantPool's internal
state I decided that subclassing is a cleaner approach. It required
significant refactoring because we had to make sure that we create
an instance of proper class when initializing the `pool` variable.
I ended up introducing `ConstantPoolManager` class which is essentially
a mutable variable that can be reset multiple times.

This change makes ClassfileParser independent from the ICodeReader and
its implementation details.