The `test-junit` Eclipse project depends on `scala-compiler`
We want to write JUnit tests for Scala compiler so we should depend
on `scala-compiler` project in `test-junit`.

Also, make dependencies of `scala-compiler` project exported so it's
enough to depend on `scala-compiler` to get all the transitive
dependencies that are needed by the compiler. Otherwise, one would
need to copy all dependencies of Scala compiler to all Eclipse projects
that depend on the compiler.