Deprecate Platform.BinaryRepr.
The 27c8266e introduced `Platform.BinaryRepr` abstract type that
was meant to abstract over binary representation of a classfile.

The abstraction was needed at the time because we had both jvm
backend and msil backend. The msil backend has been removed but
the abstraction was kept in place. The type has been deprecated
and set to be an alias for AbstractFile. The alias will be
removed later in 2.11 milestone cycle.

All references to `Platform.BinaryRepr` has been removed in
the compiler in favour of using AbstractFile directly.

If we ever need to abstract over multiple backends we should
do that based on real requirements of given backends.

This is a first step towards breaking the cyclic dependency
between Platform and SymbolLoaders.