Demonstrate forwarder detection in bytecode
Want to spot a forwarder method? Step right up, Scala
is full of them! Traits, value classes, objects, and implicit
classes leave a variety of forwarders in bytecode.

Code coverage tools and debuggers need to identify these
as compiler generated. This commit shows how the patterns
that one must look for.

In answer to an anticipated a question: these methods can't be
marked as ACC_SYNTHETIC in bytecode, as they must be visible to
Java source. Sadly, we don't have any other bits to work with in
the bytecode to mark them explicitly.