Avoid a memory leak with constructor defaults.
Namers plays host to an elaborate handoff of a tree and a Namer
from the constructor to the companion object into which the
default getters are installed.

This commit nulls out the Namer in the symbol attachment once
we're done with it to avoid holding onto that for all eternity.

We could also remove the attachment altogether, but I'm a
unwilling to touch this code more than needed right now.