Optimize ByteCodeWriters
Only run the error diagnostic for "you have a file in the way, dufus!"
if we hit an IOException.

Also routes around another layer of `isDirectory` assertions in `AbstractFile`

Saves a large number of repeated calls to file("/target/classes/com").isDirectory.

Even with the pipelined bytecode emission by GenBCode, this seems to offer a
small improvement in benchmarks.

Manual Test:

    % cat sandbox/test.scala
    package foo {
      package bar {
        class D
      }
    }

    % touch foo
    % qbin/scalac sandbox/test.scala
    sandbox/test.scala:3: error: error writing class D: ./foo/bar/D.class: ./foo is not a directory
        class D
              ^
    one error found
    % rm foo
    % mkdir foo; touch foo/bar
    % qbin/scalac sandbox/test.scala
    sandbox/test.scala:3: error: error writing class D: ./foo/bar/D.class: ./foo/bar is not a directory
        class D
              ^
    one error found