WIP SI-7090 don't leak free case class tparams into pattern pts
Regressed in 95d5554b9a2 (see also a follow up, 5708e9d73)

We were using a `tpe_*` of a polymorphic case class constructor
in place of the dangerous (kind-polymorphic) expected type of `Any`,
but didn't cleanup the floating type variables after
`inferConstructorInstance`.

In this commit,
  - add the case class type paramters to the list of symbols that
    must be existentially extrapolated
  - change the way the extrapolator is used so that it doesn't
    destructively mutate the type of the parameter symbols
    of the case class. The removed comment, "no need to clone,
    this is OUR method type", wasn't accurate (anymore?)

Without the second change, we would see crosstalk between the
two matches in:

	case class Some[a](x: a)

	object test {
	  (null: Any) match {
	    case Some(x) => x: Any
	  }

	  (null: Some[String]) match {
	    case Some(x) => x: String
	  }
	}

They would work individually, but not together.

Why didn't we spot this before? I believe that it was because the
existential extrapolation would have been a noop if we inferred
the MethodType with the original symbols, because their types could
not refer to any of the temporary GADT skolems that we were
extrapoloting over.

TODO test this theory!

TODO SI-7090 is probably a second bird that could be killed with this
     stope