SI-8085 Fix BrowserTraverser for package objects
A source file like:

    import foo.bar
    package object baz

Is parsed into:

    package <empty> {
       import foo.bar
       package baz {
          object `package`
       }
    }

A special case in Namers compensates by adjusting the owner of
`baz` to be `<root>`, rather than `<empty>`.

This wasn't being accounted for in `BrowserTraverser`, which
underpins `-srcpath`, and allows the presentation compiler to
load top level symbols from sources outside those passes as
the list of sources to compile.