SI-8133 Revert to using pre=NoPrefix in typedIdent
(Part II, see previous commit for the tests and first fix.)

Regressed in f5c336d56, a refactoring of `typedIdent`. In that
commit, an (ostensibly) accidental change arrived, equivalent to:

    - val pre1  = if (qual == EmptyTree) NoPrefix else if (sym.isTopLevel) sym.owner.thisType else qual.tpe
    + val pre1  = if (sym.isTopLevel) sym.owner.thisType else if (qual == EmptyTree) NoPrefix else qual.tpe

This commit reverts this change to `pre1`. This is an alternative
fix to the tests in the previous commit, and might help prevent
other regressions.

Some test cases needed to be updated to reflect the more compact
trees now emitted.