Trying to make -Ynooptimise redundant
By using -optimize:false instead

It doesn't quite work yet.

In:

    SCALAC_OPTS=-optimize ./test/partest test/files/run/t7852.scala

The options are:

    scalac -optimize -optimize:false ...

And the latter humbly defers to the former through the
`if (!sett.isSetByUser)` check in `enablingIfNotSetByUser`.

To get this to work, we could change partest to append, rather than
prepend, the global options, or by making `-groupSetting:false`
unconditionally falsify the options.