WIP SI-8421 Inherited members should not shadow class type param
When binding a name to a symbol, we search up through the context
chain. At each step, we first look in the current scope. If that
fails, we look for members of the prefix of the closest enclosing
context associated with a `Template` or `PackageDef`.

However, this means we will look for an inherited type member
before finding an eponymous class type parameter, which is in
the scope of an enclosing context two levels out.

TODO:

!! 1 - neg/null-unsoundness.scala                [output differs]
!! 2 - neg/volatile_no_override.scala            [expected compilation failure]