SI-8460 Refactor typedImplicit, DivergentImplicitTypeError
Organize things so that Context.openImplicits contains the full implicit
search path in the places where we create DivergentImplicitTypeError,
and capture that stack in the error. It contains all the information
we need to produce the current error message.

A subsequent commit will use this stack to improve the diagnostic message.