WIP SI-6130 substitute stray <unapply-selector>
Which leak out of extractors with dependent method types.

But, the scrutinee of the match is widened before it is bound
to intermediate values, so we still end up with a type error.

Before this patch:

    sandbox/test.scala:2: error: error during expansion of this match (this is a scalac bug).
    The underlying error was: type mismatch;
     found   : <unapply-selector>.U
     required: t.U
      (null: Any) match {
                  ^

After this patch:

    sandbox/test.scala:2: error: error during expansion of this match (this is a scalac bug).
    The underlying error was: type mismatch;
     found   : x2.U
     required: t.U
      (null: Any) match {
                  ^
    one error found

The expansion:

    {
      case <synthetic> val x1: Any = (null: Any);
      case6(){
        if (x1.isInstanceOf[Test.T])
          {
            <synthetic> val x2: Test.T = (x1.asInstanceOf[Test.T]: Test.T);
            {
              <synthetic> val o8: Option[x2.U] = Test.this.Extractor.unapply(x2);
              if (o8.isEmpty.unary_!)
                {
                  val otherExRep: t.U = o8.get;
                  matchEnd5(scala.this.Predef.println(otherExRep))
                }
              else
                case7()
            }
          }
        else
          case7()
      };
      case7(){
        matchEnd5(throw new MatchError(x1))
      };
      matchEnd5(x){
        x
      }
    }

Note that we've lost track of the symbol `t` (it was replaced
by the binder `x1`, but we didn't subtitute all references).

Furthermore, the type of `x2` has widened away the singleton
type of `x1`, it could rather be `x1.type with Test.T`.

See also: SI-7459