SI-8586 Fix rangepos crasher with defaults, poly methods
Regressed in SI-7915 / 3009a525b5

We should be deriving the position of the synthetic `Select`
from `basefun`, rather than `basefun`. In the new, enclosed
test, the difference amounts to:

    new Container().typeParamAndDefaultArg[Any]()
    `------------ basefun1 --------------'
    `----------------- basefun ---------------'

For monomorphic methods, these are one and the same, which is
why `presentation/t7915` was working. I've extended that test
to a polymorphic method to check that hyperlink resolution works.